requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
importlib-metadata==1.6.1
Markdown==3.2.2
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
zipp==3.1.0
